move isBounds method to math package and update imports where used
